Made the SpecResolver a trail and added an uncached and cached version
